FE: Refactor `allocationFlow` structure
