+++ dcapal-frontend/src/components/allocationFlow/steps/end/index.js
@@ -1,16 +1,21 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { spawn, Thread, Worker } from "threads";
-import { setAllocationFlowStep, Step } from "../../../app/appSlice";
-import { replacer, roundAmount, roundDecimals, timeout } from "../../../utils";
-import { Spinner } from "../../spinner/spinner";
+import { setAllocationFlowStep, Step } from "../../../../app/appSlice";
+import {
+  replacer,
+  roundAmount,
+  roundDecimals,
+  timeout,
+} from "../../../../utils";
+import { Spinner } from "../../../spinner/spinner";
 import {
   ACLASS,
   clearBudget,
   currentPortfolio,
   feeTypeToString,
   isWholeShares,
-} from "../portfolioStep/portfolioSlice";
+} from "../../portfolioSlice";
 import { AllocateCard } from "./allocateCard";
 import { useTranslation } from "react-i18next";
 
@@ -147,7 +152,7 @@ export const EndStep = ({ useTaxEfficient, useWholeShares }) => {
   useEffect(() => {
     const launchSolver = async () => {
       const solver = await spawn(
-        new Worker(new URL("../../../workers/solver.js", import.meta.url), {
+        new Worker(new URL("../../../../workers/solver.js", import.meta.url), {
           name: "wasm-solver-worker",
         })
       );
