+++ dcapal-frontend/src/components/allocationFlow/steps/importFile.js
@@ -2,10 +2,14 @@ import React, { useEffect, useState } from "react";
 import { useSelector } from "react-redux";
 import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
-import { setAllocationFlowStep, setPfolioFile, Step } from "../../app/appSlice";
-import { getFetcher } from "../../app/providers";
-import { timeout } from "../../utils";
-import { Spinner } from "../spinner/spinner";
+import {
+  setAllocationFlowStep,
+  setPfolioFile,
+  Step,
+} from "../../../app/appSlice";
+import { getFetcher } from "../../../app/providers";
+import { timeout } from "../../../utils";
+import { Spinner } from "../../spinner/spinner";
 import {
   ACLASS,
   FeeType,
@@ -20,7 +24,7 @@ import {
   setFeesAsset,
   setQty,
   setTargetWeight,
-} from "./portfolioStep/portfolioSlice";
+} from "../portfolioSlice";
 
 import { useTranslation } from "react-i18next";
