+++ dcapal-frontend/src/components/allocationFlow/steps/invest.js
@@ -1,17 +1,13 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useCollapse } from "react-collapsed";
-import { setAllocationFlowStep, Step } from "../../app/appSlice";
-import { InputNumber, InputNumberType } from "../core/inputNumber";
-import {
-  currentPortfolio,
-  isWholeShares,
-  setBudget,
-} from "./portfolioStep/portfolioSlice";
+import { setAllocationFlowStep, Step } from "../../../app/appSlice";
+import { InputNumber, InputNumberType } from "../../core/inputNumber";
+import { currentPortfolio, isWholeShares, setBudget } from "../portfolioSlice";
 import classNames from "classnames";
 import { Trans, useTranslation } from "react-i18next";
 import { spawn, Thread, Worker } from "threads";
-import { replacer } from "../../utils";
+import { replacer } from "../../../utils";
 
 const amtDecimals = 2;
 
@@ -62,7 +58,7 @@ export const InvestStep = ({
 
     const launchSolver = async () => {
       const solver = await spawn(
-        new Worker(new URL("../../workers/analyzer.js", import.meta.url), {
+        new Worker(new URL("../../../workers/analyzer.js", import.meta.url), {
           name: "wasm-analyzer-worker",
         })
       );
