+++ dcapal-frontend/src/components/allocationFlow/steps/portfolio/assetCard.js
@@ -2,9 +2,13 @@ import React from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useMediaQuery } from "@react-hook/media-query";
 
-import { InputNumber, InputNumberType } from "../../core/inputNumber";
-import { currentPortfolio, isWholeShares, removeAsset } from "./portfolioSlice";
-import { MEDIA_SMALL } from "../../../app/config";
+import { InputNumber, InputNumberType } from "../../../core/inputNumber";
+import {
+  currentPortfolio,
+  isWholeShares,
+  removeAsset,
+} from "../../portfolioSlice";
+import { MEDIA_SMALL } from "../../../../app/config";
 import { useCollapse } from "react-collapsed";
 import classNames from "classnames";
 import { TransactionFees } from "./transactionFees";
