+++ dcapal-frontend/src/components/allocationFlow/steps/portfolios/index.js
@@ -1,0 +1,1 @@
+import React from "react";
