+++ dcapal-frontend/src/components/exportBtn.js
@@ -5,7 +5,7 @@ import {
   aclassToString,
   currentPortfolio,
   feeTypeToString,
-} from "./allocationFlow/portfolioStep/portfolioSlice";
+} from "./allocationFlow/portfolioSlice";
 import { useTranslation } from "react-i18next";
 
 const pad = (n) => `${n}`.padStart(2, "0");
